noop is now a task for Gulp example
